Added mobility function for one of the characters
The movement function includes left and right movement and jumping, and the jumping function is divided into one section jumping and two sections gliding.
The ascending period of a jump is not smooth enough to reach the highest point in an instant, resulting in a lack of smooth connection between the first jump and the second glide.